+++ src/redux/board.slice.ts
@@ -1,7 +1,7 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
-import { getBoard } from "src/api/drello-api/board.api";
+import { getBoard } from "src/features/board/board.api";
 import { Board as innerBoard } from "src/types/board.g";
-import { Board as OuterBoard } from "src/api/drello-api/board.api";
+import { Board as OuterBoard } from "src/features/board/board.api";
 import { RootState } from "src/redux/root";
 
 interface BoardState {
