+++ src/redux/card.slice.ts
@@ -1,6 +1,6 @@
 import { createSlice } from "@reduxjs/toolkit";
 import { Card as innerCard } from "src/types/card.g";
-import { Card as OuterCard } from "src/api/drello-api/board.api";
+import { Card as OuterCard } from "src/features/board/board.api";
 import { RootState } from "src/redux/root";
 import { getBoardThunk } from "src/redux/board.slice";
